+++ packages/spor-react/src/logo/index.tsx
@@ -1,2 +1,3 @@
 export * from "./VyLogo";
 export * from "./VyLogoPride";
+export * from "./VyLogoDefault";
